@@ -22,56 +22,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/router": "^5.2.9",
-    "@covalent/core": "1.0.0",
-    "angular-async-local-storage": "^3.1.4",
+    "@angular/animations": "^7.2.1",
+    "@angular/cdk": "^7.2.1",
+    "@angular/common": "^7.2.1",
+    "@angular/compiler": "^7.2.1",
+    "@angular/core": "^7.2.1",
+    "@angular/forms": "^7.2.1",
+    "@angular/http": "^7.2.1",
+    "@angular/material": "^7.2.1",
+    "@angular/platform-browser": "^7.2.1",
+    "@angular/platform-browser-dynamic": "^7.2.1",
+    "@angular/router": "^7.2.1",
+    "@covalent/core": "2.0.0",
+    "align-yaml": "^0.1.8",
     "angular-font-awesome": "^3.1.2",
-    "core-js": "^2.5.4",
+    "buffer": "^5.2.1",
+    "core-js": "^2.6.2",
     "file-saver": "^2.0.0",
     "firacode": "^1.205.0",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "is-gzip": "^2.0.0",
-    "ng2-ace-editor": "0.3.4",
+    "js-beautify": "^1.8.9",
+    "ng2-ace-editor": "^0.3.9",
     "pako": "^1.0.8",
-    "roboto-fontface": "^0.9.0",
-    "rxjs": "^5.5.8",
+    "roboto-fontface": "^0.10.0",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "spinkit": "^1.2.5",
     "text-encoding-utf-8": "^1.0.2",
+    "tslib": "^1.9.0",
     "tsmonad": "^0.8.0",
-    "yamljs": "^0.3.0",
-    "zone.js": "^0.8.21"
+    "yaml": "^1.2.1",
+    "zone.js": "^0.8.28"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.2",
-    "@angular/compiler-cli": "^5.2.9",
-    "@types/jasmine": "2.8.6",
-    "@types/node": "9.6.1",
+    "@angular-devkit/build-angular": "^0.12.2",
+    "@angular/cli": "^7.2.1",
+    "@angular/compiler-cli": "^7.2.1",
+    "@types/jasmine": "^3.3.7",
+    "@types/node": "^10.12.18",
     "@types/text-encoding-utf-8": "^1.0.1",
-    "codelyzer": "4.2.1",
-    "http-proxy-middleware": "^0.18.0",
-    "jasmine-core": "^2.8.0",
+    "codelyzer": "^4.5.0",
+    "http-proxy-middleware": "^0.19.1",
+    "jasmine-core": "^3.3.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
+    "karma": "3.1.4",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-cli": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "node-sass": "^4.11.0",
-    "protractor": "5.3.0",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.7.2"
+    "protractor": "5.4.2",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.12.1",
+    "typescript": "3.2.4"
   }
 }
