@@ -18,7 +18,8 @@
 import {AfterViewInit, ChangeDetectorRef, Component, OnInit, ViewChild, ViewContainerRef} from "@angular/core";
 import {ActivatedRoute, Router} from "@angular/router";
 import {TdMediaService} from "@covalent/core";
-import {Observable} from "rxjs/Rx";
+import {Observable, of, from, EMPTY} from "rxjs";
+import {map, mapTo, switchMap, catchError, pluck} from "rxjs/operators";
 import {Either, Maybe} from "tsmonad";
 import {Ordering} from "./ordering";
 import {EDITOR_QUERY_NODE_PATH} from "./editor-routing.constants";
@@ -74,19 +75,24 @@ export class EditorComponent implements OnInit, AfterViewInit {
   ngOnInit(): void {
     this.zPath = this.route
       .queryParams
-      .pluck(EDITOR_QUERY_NODE_PATH)
-      .map((maybePath: string) => this.zPathService.parse(maybePath || "/"));
+      .pipe(
+        pluck(EDITOR_QUERY_NODE_PATH),
+        map((maybePath: string) => this.zPathService.parse(maybePath || "/"))
+      );
 
-    this.zNode = (<Observable<Either<Error, ZNodeWithChildren>>> this.route.data.pluck("zNodeWithChildren"))
-      .map(either =>
-        either.caseOf<Maybe<ZNodeWithChildren>>({
-          left: error => {
-            this.dialogService.showError(error.message, this.viewContainerRef);
+    this.zNode = (<Observable<Either<Error, ZNodeWithChildren>>>this.route.data)
+      .pipe(
+        pluck("zNodeWithChildren"),
+        map((either: Either<Error, ZNodeWithChildren>) =>
+          either.caseOf<Maybe<ZNodeWithChildren>>({
+            left: error => {
+              this.dialogService.showError(error.message, this.viewContainerRef);
 
-            return Maybe.nothing();
-          },
-          right: node => Maybe.just(node)
-        })
+              return Maybe.nothing();
+            },
+            right: node => Maybe.just(node)
+          })
+        )
       );
 
     this.zNode.forEach(maybeNode =>
@@ -111,16 +117,18 @@ export class EditorComponent implements OnInit, AfterViewInit {
   showSessionInfo(): void {
     this.zSessionHandler
       .getSessionInfo()
-      .switchMap((sessionInfo) => {
-          if (sessionInfo) {
-            return this.dialogService.showSessionInfo(
-              sessionInfo,
-              this.viewContainerRef
-            );
-          }
+      .pipe(
+        switchMap((sessionInfo) => {
+            if (sessionInfo) {
+              return this.dialogService.showSessionInfo(
+                sessionInfo,
+                this.viewContainerRef
+              );
+            }
 
-          return Observable.empty();
-        }
+            return EMPTY;
+          }
+        )
       )
       .subscribe();
   }
@@ -128,9 +136,11 @@ export class EditorComponent implements OnInit, AfterViewInit {
   disconnect(): void {
     this.zSessionHandler
       .getSessionInfo()
-      .switchMap(sessionInfo => this.zSessionService.close(sessionInfo))
-      .switchMap(() => this.zSessionHandler.setSessionInfo(null))
-      .catch(error => this.dialogService.showError(error, this.viewContainerRef))
+      .pipe(
+        switchMap(sessionInfo => this.zSessionService.close(sessionInfo)),
+        switchMap(() => this.zSessionHandler.setSessionInfo(null)),
+        catchError(error => this.dialogService.showError(error, this.viewContainerRef))
+      )
       .forEach(() => this.router
         .navigate(["/connect"])
         .catch(error => this.dialogService.showError(error, this.viewContainerRef))
