@@ -15,25 +15,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {Injectable} from "@angular/core";
-import {Observable, from} from "rxjs";
-import {FileReaderService} from "./file-reader.service";
-
-@Injectable()
-export class DefaultFileReaderService implements FileReaderService {
-
-  readAsText(file: File, encoding?: string): Observable<string> {
-    const promise = new Promise<string>((resolve, reject) => {
-      const reader = new FileReader();
-
-      reader.onload = () => resolve(<string> reader.result);
-      reader.onabort = ev => reject(ev);
-      reader.onerror = ev => reject(ev);
-
-      reader.readAsText(file, encoding);
-    });
-
-    return from(promise);
-  }
-}
-
+export * from "./api-response"
