@@ -16,8 +16,8 @@
  */
 
 import {Injectable} from "@angular/core";
-import "yamljs";
-import {Either} from "tsmonad";
+import {Observable, throwError, of} from "rxjs";
+import {catchError} from "rxjs/operators";
 import {Formatter} from "./formatter";
 import {ModeId} from "../../content/data/mode";
 
@@ -26,11 +26,15 @@ export class JsonFormatter extends Formatter {
 
   mode: ModeId = ModeId.Json;
 
-  format(data: string): Either<Error, string> {
-    try {
-      return Either.right(JSON.stringify(JSON.parse(data), null, 4));
-    } catch (error) {
-      return Either.left(new Error("Invalid JSON"));
-    }
+  format(data: string): Observable<string> {
+    return <Observable<string>>
+      of(<string>JSON.stringify(JSON.parse(data), null, 4))
+        .pipe(
+          catchError(err => {
+            console.log(err);
+
+            return throwError("Invalid JSON");
+          })
+        );
   }
 }
